fleshing out avx2 ops
